경북대 BE_이상희 3주차 과제(2단계) (#331)
* remove: commit으로 인한 중복 클래스 삭제

* remove: 사용하지 않는 Test 삭제

* style: 매서드 순서 변경

* refactor: ProductController -> ProductService dto추가

* refactor: Product.modify 매서드 추가

* feat: Post Product 시, 리소스 위치 반환

* refactor: ProductController <- Service dto 추가

* refactor: 클래스 이름 단수형으로 변경

* remove: commit으로 인한 중복 클래스 삭제

* refactor: WishController -> WishService dto추가

* refactor: Wish.modify 매서드 추가

* feat: owner 여부 확인 로직 추가

* feat: transactional 추가

* feat: Post Wish 시, 리소스 위치 반환

* refactor: WishController <- WishService dto추가

* remove: 사용하지 않는 매서드 삭제

* refactor: AdminController에 ProductService 적용

* feat: 연관관계 매핑

* feat: 연관관계 매핑으로 인한 코드 수정

* style: 코드 포매팅

* feat: 지연로딩 설정

* feat: fetch join으로 Wish객체 조회

+++ src/main/java/gift/wish/service/WishService.java
@@ -3,65 +3,82 @@ package gift.wish.service;
 import gift.product.domain.Product;
 import gift.product.exception.ProductNotFoundException;
 import gift.product.persistence.ProductRepository;
-import gift.wish.application.dto.request.WishRequest;
-import gift.wish.application.dto.response.WishResponse;
+import gift.user.domain.User;
+import gift.user.persistence.UserRepository;
 import gift.wish.domain.Wish;
-import gift.wish.exception.WishCanNotModifyException;
 import gift.wish.exception.WishNotFoundException;
 import gift.wish.persistence.WishRepository;
+import gift.wish.service.dto.WishInfo;
+import gift.wish.service.dto.WishParam;
 import java.util.List;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class WishService {
     private final WishRepository wishRepository;
     private final ProductRepository productRepository;
+    private final UserRepository userRepository;
 
-    public WishService(WishRepository wishRepository, ProductRepository productRepository) {
+    public WishService(WishRepository wishRepository,
+                       ProductRepository productRepository,
+                       UserRepository userRepository
+    ) {
         this.wishRepository = wishRepository;
         this.productRepository = productRepository;
+        this.userRepository = userRepository;
     }
 
-    public void saveWish(WishRequest wishRequest, final Long userId) {
+    public Long saveWish(WishParam wishRequest) {
         Product product = productRepository.findById(wishRequest.productId())
                 .orElseThrow(() -> ProductNotFoundException.of(wishRequest.productId()));
+        User user = userRepository.getReferenceById(wishRequest.userId());
 
-        Wish wish = wishRequest.toModel(userId);
+        Wish wish = new Wish(wishRequest.amount(), product, user);
         wishRepository.save(wish);
+        return wish.getId();
     }
 
-    public void updateWish(Long wishId, WishRequest wishRequest, final Long userId) {
-        Wish wish = wishRepository.findByIdAndUserId(wishId, userId)
+    @Transactional
+    public void updateWish(WishParam wishRequest, final Long wishId) {
+        Wish wish = wishRepository.findWishByIdWithUserAndProduct(wishId)
                 .orElseThrow(() -> WishNotFoundException.of(wishId));
 
-        if (!wish.getProductId().equals(wishRequest.productId())) {
-            throw new WishCanNotModifyException();
-        }
+        User user = userRepository.getReferenceById(wishRequest.userId());
+        Product product = productRepository.getReferenceById(wishRequest.productId());
 
-        Wish newWish = wishRequest.toModel(wishId, userId);
-        wishRepository.save(newWish);
+        wish.modify(wishRequest.amount(), product, user);
     }
 
-    public List<WishResponse> getWishList(final Long userId) {
-        List<Wish> wishes = wishRepository.findWishesByUserId(userId);
+    @Transactional(readOnly = true)
+    public List<WishInfo> getWishList(final Long userId) {
+        List<Wish> wishes = wishRepository.findWishesByUserIdWithUserAndProduct(userId);
 
-        List<WishResponse> responses = wishes.stream()
-                .map(WishResponse::fromModel)
+        var responses = wishes.stream()
+                .map(WishInfo::from)
                 .toList();
 
         return responses;
     }
 
-    public WishResponse getWish(final Long wishId, final Long userId) {
-        Wish wish = wishRepository.findByIdAndUserId(wishId, userId)
+    @Transactional(readOnly = true)
+    public WishInfo getWish(final Long wishId, final Long userId) {
+        Wish wish = wishRepository.findById(wishId)
                 .orElseThrow(() -> WishNotFoundException.of(wishId));
+        if (!wish.isOwner(userId)) {
+            throw WishNotFoundException.of(wishId);
+        }
 
-        return WishResponse.fromModel(wish);
+        return WishInfo.from(wish);
     }
 
+    @Transactional
     public void deleteWish(final Long wishId, final Long userId) {
-        Wish wish = wishRepository.findByIdAndUserId(wishId, userId)
+        Wish wish = wishRepository.findById(wishId)
                 .orElseThrow(() -> WishNotFoundException.of(wishId));
+        if (!wish.isOwner(userId)) {
+            throw WishNotFoundException.of(wishId);
+        }
 
         wishRepository.delete(wish);
     }
